Rename actionHandlers package
