+++ ci-droid-tasks-consumer-services/src/main/java/com/societegenerale/cidroid/tasks/consumer/services/PullRequestEventService.java
@@ -1,6 +1,6 @@
 package com.societegenerale.cidroid.tasks.consumer.services;
 
-import com.societegenerale.cidroid.tasks.consumer.services.actionHandlers.PullRequestEventHandler;
+import com.societegenerale.cidroid.tasks.consumer.services.eventhandlers.PullRequestEventHandler;
 import com.societegenerale.cidroid.tasks.consumer.services.model.github.PullRequestEvent;
 import com.societegenerale.cidroid.tasks.consumer.services.monitoring.Event;
 import com.societegenerale.cidroid.tasks.consumer.services.monitoring.MonitoringEvents;
